Add files via upload
Write a program to store and manipulate students records. More details are as follows.
1. Store following students information.
	a. Name, roll no, age, address, email, mobile number, program name (ex BTech/MTech) and semester
	b. two compulsory subjects and two elective subjects opted for the semester
	c. list of fifteen subjects, max capacity, and semester (this value describes which semester students can opt for it)
2. Write functions that allow user to store these information
3. Write a function that allows a student to register in a particular sememster and also allots compulsory courses if registration is successful
4. Write functions that allow a student to opt for a subject (elective)
5. Write functions to modify, search and display records of students
Note: max number of students is not known in advance, subjects have a max capacity and a semester, students can not opt a subject if they are not in that semester or higher. As soon as a student registers, compulsory courses are allocated and electives are set to not yet allotted.
Hint: use the concepts of structure, array, searching, and dynamic memory allocation.